% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_trip.R
\name{agg_trip}
\alias{agg_trip}
\title{aggregate NDS ping data to shift or trips new datasets.}
\usage{
agg_trip(data, trip_id)
}
\arguments{
\item{data}{The data to be passed on.}

\item{trip_id}{The ID column generated by the `segment()` function.}

\item{threhold}{The threshold you want to separate the ping (in minutes). It is recommended to use 30 minutes to separate into trips, and 8*60 minutes to separate into shifts.}

\item{time_diff}{The time difference between the nearest two pings.}
}
\description{
This function produce new datasets with starting and ending variables, which may include date and time, latitude, longitude.
}
\examples{
`agg_trip(data = dat, trip_id = id)`
}
\keyword{aggregate}
