% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDS-package.R
\docType{package}
\name{NDS-package}
\alias{NDS}
\alias{NDS-package}
\title{NDS: An R package for manipulating Naturalistic Driving Datasets (NDS).}
\description{
An R package for manipulating Naturalistic Driving Datasets (NDS).
}
\author{
\strong{Maintainer}: Miao Cai \email{miao.cai@slu.edu} (0000-0003-0170-6905)

}
\keyword{(NDS)}
\keyword{driving}
\keyword{naturalistic}
\keyword{studies}
