% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment}
\alias{segment}
\title{`segment` NDS ping data to shifts and trips}
\usage{
segment(speed, threshold, time_diff)
}
\arguments{
\item{speed}{Real time speed of the ping}

\item{time_diff}{The time difference between the nearest two pings.}

\item{threhold}{The threshold you want to separate the ping (in minutes). It is recommended to use 30 minutes to separate into trips, and 8*60 minutes to separate into shifts.}
}
\description{
This function segments NDS ping data to shifts and trips
}
\examples{
`segment(dat$speed, 30, d$time_diff)`
}
\keyword{NDS}
\keyword{ping}
\keyword{separate}
\keyword{shift}
\keyword{trip}
