% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDS-package.R
\docType{package}
\name{NDS-package}
\alias{NDS}
\alias{NDS-package}
\title{NDS: a toolkit for handling NDS datasets.}
\description{
NDS provides high-performance data separation,
aggregation, reorganize, and fuzzy joining toolkits to
handle high-resolution naturalistic driving datasets.
}
\details{
\itemize{
\item Aggregate large real-time ping data to trips or shifts.
\item Separate trips or shifts to equal-length time intervals.
\item Merge Safety Critical-Events with trips, shifts, and time intervals.
}

To learn more about NDS, you can start with the vignettes:
`browseVignettes(package = "NDS")`
}
\author{
\strong{Maintainer}: Miao Cai \email{miao.cai@slu.edu} (0000-0003-0170-6905)

}
